Also move gemm related registration files
